requirements.txt
openai
faiss-cpu
streamlit
pandas
python-dotenv
pyyaml
tiktoken
sentence-transformers
